backend/requirements.txt

uvicorn[standard]
pypdf
chromadb
fastapi
sentence-transformers
ollama
python-multipart
python-dotenv
